Chore: move root scripts into scripts/; update paths and npm scripts
