+++ scripts/get-auth-token.js
