+++ scripts/set-admin-claims.js
@@ -2,7 +2,7 @@ const admin = require('firebase-admin');
 const path = require('path');
 
 // Initialize Firebase Admin SDK
-const serviceAccount = require('./server/firebase-stage-credentials.json');
+const serviceAccount = require('../server/firebase-stage-credentials.json');
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount),
 });
